+++ Editor/Artifice_CustomAttributeDrawers/CustomAttributeDrawer_EnableIfAttribute/Artifice_CustomAttributeDrawer_EnableIfAttribute.cs
@@ -1,69 +1,96 @@
+using System;
 using System.Linq;
 using ArtificeToolkit.Attributes;
 using UnityEditor;
 using UnityEditor.UIElements;
-using UnityEngine;
 using UnityEngine.UIElements;
 
 namespace ArtificeToolkit.Editor.Artifice_CustomAttributeDrawers.CustomAttributeDrawer_EnableIfAttribute
 {
-    /// <summary> Custom VisualAttribute drawer for <see cref="EnableIfAttribute"/> </summary>
     [Artifice_CustomAttributeDrawer(typeof(EnableIfAttribute))]
-    public class Artifice_CustomAttributeDrawer_EnableIfAttribute : Artifice_CustomAttributeDrawer
+    public class Artifice_CustomAttributeDrawer_EnableIfAttribute : Artifice_CustomAttributeDrawer, IArtifice_ShouldIncludeInValidation
     {
-        #region FIELDS
-
         private EnableIfAttribute _attribute;
         private SerializedProperty _trackedProperty;
         private VisualElement _targetElem;
-       
-        #endregion
-        
-        /* Get reference to VisualElement to target */
+        private bool _isReflected;
+
         public override VisualElement OnWrapGUI(SerializedProperty property, VisualElement root)
         {
             _targetElem = root;
-            
             _attribute = (EnableIfAttribute)Attribute;
-            
-            // Set Data tracked property
+
+            // Try find property (serialized or reflected)
             _trackedProperty = property.FindPropertyInSameScope(_attribute.PropertyName);
             if (_trackedProperty == null)
-                Debug.LogWarning("Cannot find property with name " + _attribute.PropertyName);
+                _isReflected = true;
+        
+            // Setup tracking depending on type
+            // 1) Serialized Property
+            // 2) Reflection
+            if (_isReflected == false)
+            {
+                // Use Unity’s built-in tracker for serialized fields
+                UpdateRootVisibility(_trackedProperty);
+                var trackerElement = new VisualElement { name = "Tracker Element" };
+                _targetElem.Add(trackerElement);
+                trackerElement.TrackPropertyValue(_trackedProperty, UpdateRootVisibility);
+            }
+            else
+            {
+                // Execute once and every 400 ms
+                EvaluateReflectedCondition(property);
+                root.schedule.Execute(() => EvaluateReflectedCondition(property))
+                    .Every(400);
+            }
 
-            UpdateRootVisibility(_trackedProperty);
-            
-            var trackerElement = new VisualElement();
-            trackerElement.name = "Tracker Element";
-            trackerElement.tooltip = "Used only for TrackPropertyValue method";
-            _targetElem.Add(trackerElement);
-            trackerElement.TrackPropertyValue(_trackedProperty, UpdateRootVisibility);
-            
             return _targetElem;
         }
+
+        private void EvaluateReflectedCondition(SerializedProperty property)
+        {
+            var target = property.serializedObject.targetObject;
+            if (target == null)
+                return;
+
+            property.FindReflectedPropertyInSameScope(_attribute.PropertyName, out var returnValue);
         
-        /* Executes logic on changing visibility */
+            var shouldShow = _attribute.Values.Any(value => Artifice_Utilities.AreEqual(returnValue, value));
+            SetVisibility(shouldShow);
+        }
+
         private void UpdateRootVisibility(SerializedProperty trackedProperty)
         {
             var trackedValue = trackedProperty.GetTarget<object>();
-            
-            if(_attribute.Values.Any(value => Artifice_Utilities.AreEqual(trackedValue, value)))
+            var shouldShow = _attribute.Values.Any(v => Artifice_Utilities.AreEqual(trackedValue, v));
+            SetVisibility(shouldShow);
+        }
+
+        private void SetVisibility(bool visible)
+        {
+            if (visible)
                 _targetElem.RemoveFromClassList("hide");
             else
                 _targetElem.AddToClassList("hide");
         }
-
-        /// <summary> This method is called by Validator_Modules to be able to skip properties which are not enabled. This is not the most generic approach, but it will serve as a fine test to see if this is the desired behaviour. </summary>
-        /// <remarks> In the future and if needed, it will probably take the form of a interface. If it becomes an interface, it will also cost extra performance to parse all the attributes and type check them. </remarks> 
-        public static bool ShouldIncludeInValidation(SerializedProperty property, EnableIfAttribute attribute)
+    
+        public bool ShouldIncludeInValidation(SerializedProperty property, CustomAttribute customAttribute)
         {
-            // Set Data tracked property
+            var attribute = customAttribute as EnableIfAttribute;
+            if (attribute == null)
+                throw new ArgumentException();
+        
             var trackedProperty = property.FindPropertyInSameScope(attribute.PropertyName);
-            if (trackedProperty == null)
-                return false;
-            
-            var trackedValue = trackedProperty.GetTarget<object>();
-            return attribute.Values.Any(value => Artifice_Utilities.AreEqual(trackedValue, value));
+            if (trackedProperty != null)
+            {
+                var targetValue = trackedProperty.GetTarget<object>();
+                return attribute.Values.Any(value => Artifice_Utilities.AreEqual(targetValue, value));
+            }
+            else
+            {
+                property.FindReflectedPropertyInSameScope(attribute.PropertyName, out var returnValue);
+                return attribute.Values.Any(v => Artifice_Utilities.AreEqual(returnValue, v));
+            }    
         }
     }
 }