+++ Editor/Artifice_Validator/Artifice_ValidatorExtensions.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using ArtificeToolkit.Attributes;
-using ArtificeToolkit.Editor.Artifice_CustomAttributeDrawers.CustomAttributeDrawer_EnableIfAttribute;
 using ArtificeToolkit.Editor.Artifice_CustomAttributeDrawers.CustomAttributeDrawer_Validators;
 using UnityEditor;
 using UnityEditor.SceneManagement;
@@ -12,9 +11,22 @@ namespace ArtificeToolkit.Editor
 {
     public static class Artifice_ValidatorExtensions
     {
-        /// <summary> Fills in-parameter list with logs found in property </summary>
-        /// <remarks>Passing the List as a parameter as a minor optimization to avoid instantiating the list on each call of GenerateValidatorLogs.</remarks>
-        public static void GenerateValidatorLogs(SerializedProperty property, List<Artifice_Validator.ValidatorLog> logs, Type validatorType)
+        /// <summary> Fills in-parameter list with logs found among the custom attribtues in the property </summary>
+        /// <remarks>
+        /// This overload uses <see cref="SerializedProperty.GetCustomAttributes"/> to fetch attributes.
+        /// Passing the List as a parameter as a minor optimization to avoid instantiating the list on each call of GenerateValidatorLogs.
+        /// </remarks>
+        public static void GenerateValidatorLogs(SerializedProperty property,
+            List<Artifice_Validator.ValidatorLog> logs, Type validatorType)
+        {
+            GenerateValidatorLogs(property, p => p.GetCustomAttributes(), logs, validatorType);
+        }
+        
+        /// <summary> Fills in-parameter list with logs found in the property using the attribute provider </summary>
+        /// <remarks> Passing the List as a parameter as a minor optimization to avoid instantiating the list on each call of GenerateValidatorLogs. </remarks>
+        public static void GenerateValidatorLogs(SerializedProperty property, 
+            Func<SerializedProperty, IEnumerable<CustomAttribute>> attributeProvider, 
+            List<Artifice_Validator.ValidatorLog> logs, Type validatorType)
         {
             if (property.IsArray())
             {
@@ -23,7 +35,7 @@ namespace ArtificeToolkit.Editor
                 var childrenCustomAttributes = new List<CustomAttribute>();
 
                 // Get property attributes and parse-split them
-                var attributes = property.GetCustomAttributes();
+                var attributes = attributeProvider(property);
                 if (attributes != null)
                     foreach (var attribute in attributes)
                         if (attribute is IArtifice_ArrayAppliedAttribute)
@@ -42,25 +54,30 @@ namespace ArtificeToolkit.Editor
             else
             {
                 // Check property if its valid for stuff
-                var customAttributes = property.GetCustomAttributes();
+                var customAttributes = attributeProvider(property);
                 if (customAttributes != null)
                     GenerateValidatorLogs(property, customAttributes.ToList(), logs, validatorType);
             }
         }
 
         /// <summary> Fills in-parameter list with logs found in property for specific parameterized attributes</summary>
-        private static void GenerateValidatorLogs(SerializedProperty property, List<CustomAttribute> customAttributes, List<Artifice_Validator.ValidatorLog> logs, Type validatorType)
+        public static void GenerateValidatorLogs(SerializedProperty property, List<CustomAttribute> customAttributes, List<Artifice_Validator.ValidatorLog> logs, Type validatorType)
         {
             var validatorAttributes = customAttributes.Where(attribute => attribute is ValidatorAttribute).ToList();
             foreach (var validatorAttribute in validatorAttributes)
             {
                 // Get drawer and cast to validator drawer.
-                var drawer =
-                    Artifice_Utilities.GetDrawerInstancesMap()[validatorAttribute.GetType()] as
-                        Artifice_CustomAttributeDrawer_Validator_BASE;
+                var drawerInstancesMap = Artifice_Utilities.GetDrawerInstancesMap();
+                if (drawerInstancesMap.ContainsKey(validatorAttribute.GetType()) == false)
+                {
+                    Artifice_Utilities.LogError($"Could not find validator drawer for <b>{validatorAttribute.GetType().Name}</b>.");
+                    continue;
+                }
+                
+                var drawer =drawerInstancesMap[validatorAttribute.GetType()] as Artifice_CustomAttributeDrawer_Validator_BASE;
                 if (drawer == null)
                 {
-                    Debug.LogWarning($"Could not find drawer for validator type {validatorAttribute.GetType().Name}");
+                    Artifice_Utilities.LogError($"Could not find drawer for validator type {validatorAttribute.GetType().Name}.");
                     continue;
                 }
 
@@ -102,15 +119,32 @@ namespace ArtificeToolkit.Editor
                 }
             }
         }
-
+        
         /// <summary> Returns true if property should be taken into consideration in the validation or not. </summary>
         public static bool ShouldValidateProperty(SerializedProperty property)
         {
-            // Check if property is under enable if => false. In that case skip that property and its children.
-            var customAttributes = property.GetCustomAttributes().ToList();
-            var attribute = customAttributes.Find(attribute => attribute is EnableIfAttribute);
-            if (attribute is EnableIfAttribute enableIfAttribute)
-                return Artifice_CustomAttributeDrawer_EnableIfAttribute.ShouldIncludeInValidation(property, enableIfAttribute);
+            // Check if property is under an occasional validation which evaluates to false. In that case skip that property and its children.
+            return ShouldValidateProperty(property, property.GetCustomAttributes());
+        }
+        
+        /// <summary> Returns true if property should be taken into consideration in the validation or not. </summary>
+        public static bool ShouldValidateProperty(SerializedProperty property, CustomAttribute[] customAttributes)
+        {
+            foreach (var attribute in customAttributes)
+            {
+                if (attribute is not IArtifice_RequiresCheckForValidationInclusion)
+                    continue;
+             
+                // Get drawer, cast to IArtifice_ShouldIncludeInValidation and run check.
+                var customAttributeDrawer = Artifice_Utilities.GetDrawerInstancesMap()[attribute.GetType()];
+                var shouldIncludeInValidation = customAttributeDrawer as IArtifice_ShouldIncludeInValidation;
+                if (shouldIncludeInValidation != null)
+                    return shouldIncludeInValidation.ShouldIncludeInValidation(property, attribute);
+                else
+                    Debug.LogWarning($"<color=yellow>[ArtificeToolkit]</color> {attribute.GetType().Name} implements " +
+                                     $"{nameof(IArtifice_RequiresCheckForValidationInclusion)} but its corresponding " +
+                                     $"drawer does not implement {nameof(IArtifice_ShouldIncludeInValidation)}");
+            }
 
             return true;
         }